+++ uui/src/lib/Paper/Overlay.types.d.ts
@@ -1,0 +1,5 @@
+import type { BoxProps } from '../Box/box.types';
+
+export interface OverlayProps<T extends EventTarget> extends BoxProps<T> {
+  elevation?: number;
+}