+++ uui/src/lib/Theme/theme.svelte
@@ -9,6 +9,8 @@
   import { updateThemeStyle } from '../shared/theme/config/theme-config.utils';
   import { UUI_PALETTES } from '../shared/theme/palette/default-palettes';
   import type { Palettes } from '../shared/theme/palette/palette.types';
+  import { currentTheme } from '../shared/theme/store/theme-store';
+  import { getPreferredTheme } from '../shared/utils/getPreferredTheme';
   import '../shared/theme/default-colors/colors.css';
   import { onMount } from 'svelte';
 
@@ -19,10 +21,14 @@
   // Check for light theme otherwise use dark
   // Shared is always applied even when its the only theme.
   export let theme = palettes['light'] ? 'light' : 'dark';
+  // Find and use preferred theme if true
+  export let preferredTheme = true;
 
   // Locals
   // Used to ensure component is mounted.
   let pageDocument: Document;
+  // used for the observed theme value from theme store
+  let watchedTheme: string;
   // Convert palettes to map with key, ThemeVars.
   const paletteMap = createPaletteMap(mergePalettes(palettes));
   // Create Typography Config
@@ -30,6 +36,26 @@
 
   // merge and set BreakPoint Context
   BREAKPOINT_CONTEXT.setBreakpoints({ ...UUI_BREAKPOINTS, ...breakpoints });
+
+  // Lifecycle
+  onMount(() => {
+    pageDocument = document;
+
+    // If preferredTheme is true, get theme from window and set in the theme store
+    // else set theme store to theme passed as prop
+    const themePreference = preferredTheme && getPreferredTheme(window);
+    if (themePreference) {
+      currentTheme.set(themePreference);
+    } else {
+      currentTheme.set(theme);
+    }
+
+    // Needs to be inside onMount to ensure subscribe happens
+    // after theme store is set to either preferred theme or theme passed as prop
+    currentTheme.subscribe((value) => {
+      watchedTheme = value;
+    });
+  });
 
   // Watch theme and apply when changed.
   $: updateThemeStyle(
@@ -39,7 +65,7 @@
         // Apply shared palette first as base
         ...(paletteMap.get('shared') || {}),
         // Apply chosen theme palette
-        ...(paletteMap.get(theme) || {}),
+        ...(paletteMap.get(watchedTheme) || {}),
         // Apply the typography vars
         ...typographyConfig.vars
       },
@@ -48,11 +74,6 @@
     },
     pageDocument
   );
-
-  // Lifecycle
-  onMount(() => {
-    pageDocument = document;
-  });
 </script>
 
 <!--