+++ website/src/lib/ThemeToggle.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
-  import { Button } from '@uui';
-  import { currentTheme } from './theme/theme-store';
+  import { Button, currentTheme } from '@uui';
+
   let currentThemeValue: string;
   currentTheme.subscribe((val) => {
     currentThemeValue = val;
@@ -16,5 +16,7 @@
   on:click={toggleTheme}
   variant="raised"
   color="secondary"
-  style="display: block; margin-left: auto; margin-right: auto;">Switch Theme</Button
+  style="display: block; margin-left: auto; margin-right: auto;"
 >
+  Switch Theme
+</Button>